Step1X-3D/requirements.txt
datasets==2.19.0
diffusers==0.32.2
einops==0.8.0
huggingface-hub==0.26.2
imageio==2.34.1
jaxtyping==0.2.28
joblib==1.4.0
lightning-utilities==0.11.2
matplotlib==3.8.4
numpy==1.26.4
omegaconf==2.3.0
opencv-python-headless==4.10.0.84
pandas==2.2.2
pillow==10.3.0
plyfile==1.0.3
PyMCubes==0.1.4
pyparsing==3.1.2
pytorch-lightning==2.2.4
PyYAML==6.0.1
safetensors==0.4.3
scikit-image==0.23.2
scipy==1.13.0
tensorboard==2.16.2
tensorboardX==2.6.2.2
timm==0.9.16
tokenizers==0.21.0
tqdm==4.66.2
transformers==4.48.0
trimesh==4.3.2
spaces==0.28.3
accelerate==1.5.2
rembg==2.0.65
gradio==5.5.0
wandb==0.18.6
deepspeed==0.16.4
sageattention==1.0.6
mosaicml-streaming==0.11.0
easydict==1.13
open3d==0.19.0
prodigyopt==1.1.2 
peft==0.15.1
sentencepiece==0.2.0
pymeshlab==2023.12.post3
onnxruntime==1.21.0
bs4==0.0.2
xatlas==0.0.10
pybind11==2.13.6
pygltflib==1.16.4
kornia==0.8.0
cupy-cuda12x==13.4.1
git+https://github.com/NVlabs/nvdiffrast.git
git+https://github.com/facebookresearch/pytorch3d.git@stable
